By chasing one bug we reached another bug, and currently the code is not in a good place (needs either reverting or fixing).
Explanation: Hybrid monotonicities create problems in the lower
bound tree. Namely, if a vertex is present in the tree with packing
{10,6,9} and last item 6 (monotonicity 5), then because of the
"lowest sendable" hashing, it was hashed as {10,6,9} lowest sendable 1.

However, if we then switch to monotonicity 4, the situation
{10,6,9} last item 6 will be store as {10,6,9} lowest sendable 2,
and creates a new vertex, which is then a problem for a lower bound.

A proposed solution is to hash the tree only by packing and last item
-- this will solve the issue, because mono 4 will only see vertex of
mono 5 if it is completed, which means we can link to it freely.

However the solution is not yet there.